api/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic-settings==2.1.0
scalar-fastapi==1.0.3
azure-cosmos==4.5.1
